independant validation
