--- packages/neon-api/src/transationBuilder/index.ts
@@ -1,1 +1,0 @@
-export * from "./TransactionBuilder";
